requirements.txt
streamlit>=1.20.0
llama-index>=0.10.41
llama-index-llms-google-genai>=0.1.5
youtube-transcript-api>=0.6.2
python-dotenv>=1.0.0